+++ tests/blockchain/eth/contracts/contracts/proxy/ContractV2.sol
@@ -12,41 +12,41 @@ contract ContractV2 is ContractInterface, Upgradeable {
 
     event EventV2(uint8 value);
 
-    uint public storageValue;
+    uint256 public storageValue;
     string public dynamicallySizedValue;
-    uint[] public arrayValues;
-    mapping (uint => uint) mappingValues;
-    uint[] public mappingIndices;
+    uint256[] public arrayValues;
+    mapping (uint256 => uint256) mappingValues;
+    uint256[] public mappingIndices;
 
     struct Structure1 {
-        uint value;
-        uint[] arrayValues;
-//        uint valueToCheck; // rollback does not work
+        uint256 value;
+        uint256[] arrayValues;
+//        uint256 valueToCheck; // rollback does not work
     }
     Structure1[] public arrayStructures;
 
     struct Structure2 {
-        uint value;
-        uint[] arrayValues;
-        uint valueToCheck;
+        uint256 value;
+        uint256[] arrayValues;
+        uint256 valueToCheck;
     }
-    mapping (uint => Structure2) public mappingStructures;
-    uint public mappingStructuresLength;
+    mapping (uint256 => Structure2) public mappingStructures;
+    uint256 public mappingStructuresLength;
 
-    uint public storageValueToCheck;
+    uint256 public storageValueToCheck;
 
-    constructor(uint _storageValueToCheck) public {
+    constructor(uint256 _storageValueToCheck) public {
         storageValueToCheck = _storageValueToCheck;
     }
 
-    function returnValue() public pure returns (uint) {
+    function returnValue() public pure returns (uint256) {
         return 20;
     }
 
-    function setStorageValue(uint _value) public {
+    function setStorageValue(uint256 _value) public {
         storageValue = _value * 2;
     }
-    function getStorageValue() public view returns (uint) {
+    function getStorageValue() public view returns (uint256) {
         return storageValue;
     }
 
@@ -57,79 +57,79 @@ contract ContractV2 is ContractInterface, Upgradeable {
         return dynamicallySizedValue;
     }
 
-    function pushArrayValue(uint _value) public {
+    function pushArrayValue(uint256 _value) public {
         arrayValues.push(2 * _value);
     }
-    function getArrayValue(uint _index) public view returns (uint) {
+    function getArrayValue(uint256 _index) public view returns (uint256) {
         return arrayValues[_index];
     }
-    function getArrayValueLength() public view returns (uint) {
+    function getArrayValueLength() public view returns (uint256) {
         return arrayValues.length;
     }
 
-    function setMappingValue(uint _index, uint _value) public {
+    function setMappingValue(uint256 _index, uint256 _value) public {
         mappingIndices.push(_index);
         mappingValues[_index] = _value;
     }
-    function getMappingValue(uint _index) public view returns (uint) {
+    function getMappingValue(uint256 _index) public view returns (uint256) {
         return mappingValues[_index];
     }
 
-    function getStructureLength1() public view returns (uint) {
+    function getStructureLength1() public view returns (uint256) {
         return arrayStructures.length;
     }
-    function pushStructureValue1(uint _value) public {
+    function pushStructureValue1(uint256 _value) public {
         Structure1 memory structure;
         arrayStructures.push(structure);
         arrayStructures[arrayStructures.length - 1].value = _value;
     }
-    function getStructureValue1(uint _index) public view returns (uint) {
+    function getStructureValue1(uint256 _index) public view returns (uint256) {
         return arrayStructures[_index].value;
     }
-    function getStructureArrayLength1(uint _index) public view returns (uint) {
+    function getStructureArrayLength1(uint256 _index) public view returns (uint256) {
         return arrayStructures[_index].arrayValues.length;
     }
-    function pushStructureArrayValue1(uint _index, uint _value) public {
+    function pushStructureArrayValue1(uint256 _index, uint256 _value) public {
         arrayStructures[_index].arrayValues.push(_value);
     }
-    function getStructureArrayValue1(uint _index, uint _arrayIndex) public view returns (uint) {
+    function getStructureArrayValue1(uint256 _index, uint256 _arrayIndex) public view returns (uint256) {
         return arrayStructures[_index].arrayValues[_arrayIndex];
     }
 
-    function getStructureLength2() public view returns (uint) {
+    function getStructureLength2() public view returns (uint256) {
         return mappingStructuresLength;
     }
-    function pushStructureValue2(uint _value) public {
+    function pushStructureValue2(uint256 _value) public {
         mappingStructuresLength++;
         Structure2 memory structure;
         mappingStructures[mappingStructuresLength - 1] = structure;
         mappingStructures[mappingStructuresLength - 1].value = _value;
     }
-    function getStructureValue2(uint _index) public view returns (uint) {
+    function getStructureValue2(uint256 _index) public view returns (uint256) {
         return mappingStructures[_index].value;
     }
-    function getStructureArrayLength2(uint _index) public view returns (uint) {
+    function getStructureArrayLength2(uint256 _index) public view returns (uint256) {
         return mappingStructures[_index].arrayValues.length;
     }
-    function pushStructureArrayValue2(uint _index, uint _value) public {
+    function pushStructureArrayValue2(uint256 _index, uint256 _value) public {
         mappingStructures[_index].arrayValues.push(_value);
     }
-    function getStructureArrayValue2(uint _index, uint _arrayIndex) public view returns (uint) {
+    function getStructureArrayValue2(uint256 _index, uint256 _arrayIndex) public view returns (uint256) {
         return mappingStructures[_index].arrayValues[_arrayIndex];
     }
 
-    function setStructureValueToCheck2(uint _index, uint _value) public {
+    function setStructureValueToCheck2(uint256 _index, uint256 _value) public {
         mappingStructures[_index].valueToCheck = _value;
     }
-    function getStructureValueToCheck2(uint _index) public view returns (uint) {
+    function getStructureValueToCheck2(uint256 _index) public view returns (uint256) {
         return mappingStructures[_index].valueToCheck;
     }
 
-    function getStructure1ArrayValues(uint _index) public view returns (uint[] memory) {
+    function getStructure1ArrayValues(uint256 _index) public view returns (uint256[] memory) {
         return arrayStructures[_index].arrayValues;
     }
 
-    function getStructure2ArrayValues(uint _index) public view returns (uint[] memory) {
+    function getStructure2ArrayValues(uint256 _index) public view returns (uint256[] memory) {
         return mappingStructures[_index].arrayValues;
     }
 
@@ -140,41 +140,41 @@ contract ContractV2 is ContractInterface, Upgradeable {
             keccak256(value) == keccak256(bytes(dynamicallySizedValue)));
 
         require(delegateGet(_testTarget, "getArrayValueLength()") == arrayValues.length);
-        for (uint i = 0; i < arrayValues.length; i++) {
+        for (uint256 i = 0; i < arrayValues.length; i++) {
             require(delegateGet(_testTarget, "getArrayValue(uint256)", bytes32(i)) == arrayValues[i]);
         }
-        for (uint i = 0; i < mappingIndices.length; i++) {
-            uint index = mappingIndices[i];
+        for (uint256 i = 0; i < mappingIndices.length; i++) {
+            uint256 index = mappingIndices[i];
             require(delegateGet(_testTarget, "getMappingValue(uint256)", bytes32(index)) ==
                 mappingValues[index]);
         }
 
-        require(uint(delegateGet(_testTarget, "getStructureLength1()")) == arrayStructures.length);
-        for (uint i = 0; i < arrayStructures.length; i++) {
+        require(uint256(delegateGet(_testTarget, "getStructureLength1()")) == arrayStructures.length);
+        for (uint256 i = 0; i < arrayStructures.length; i++) {
             Structure1 memory structure1 = delegateGetStructure1(_testTarget, "arrayStructures(uint256)", bytes32(i));
             require(structure1.value == arrayStructures[i].value);
 
             bytes32[] memory values = delegateGetArray(_testTarget, "getStructure1ArrayValues(uint256)", bytes32(i));
             require(values.length == arrayStructures[i].arrayValues.length);
-            for (uint j = 0; j < arrayStructures[i].arrayValues.length; j++) {
-                require(uint(values[j]) == arrayStructures[i].arrayValues[j]);
+            for (uint256 j = 0; j < arrayStructures[i].arrayValues.length; j++) {
+                require(uint256(values[j]) == arrayStructures[i].arrayValues[j]);
             }
         }
 
-        require(uint(delegateGet(_testTarget, "getStructureLength2()")) == mappingStructuresLength);
-        for (uint i = 0; i < mappingStructuresLength; i++) {
+        require(uint256(delegateGet(_testTarget, "getStructureLength2()")) == mappingStructuresLength);
+        for (uint256 i = 0; i < mappingStructuresLength; i++) {
             Structure2 memory structure2 = delegateGetStructure2(_testTarget, "mappingStructures(uint256)", bytes32(i));
             require(structure2.value == mappingStructures[i].value);
             require(structure2.valueToCheck == mappingStructures[i].valueToCheck);
 
             bytes32[] memory values = delegateGetArray(_testTarget, "getStructure2ArrayValues(uint256)", bytes32(i));
             require(values.length == mappingStructures[i].arrayValues.length);
-            for (uint j = 0; j < mappingStructures[i].arrayValues.length; j++) {
-                require(uint(values[j]) == mappingStructures[i].arrayValues[j]);
+            for (uint256 j = 0; j < mappingStructures[i].arrayValues.length; j++) {
+                require(uint256(values[j]) == mappingStructures[i].arrayValues[j]);
             }
         }
 
-        require(uint(delegateGet(_testTarget, "storageValueToCheck()")) == storageValueToCheck);
+        require(uint256(delegateGet(_testTarget, "storageValueToCheck()")) == storageValueToCheck);
     }
 
     function delegateGetStructure1(address _target, string memory _signature, bytes32 _argument)
