+++ tests/blockchain/eth/contracts/contracts/proxy/ContractV3.sol
@@ -6,15 +6,15 @@ import "./ContractV2.sol";
 
 contract ContractV3 is ContractV2 {
 
-    uint public anotherStorageValue;
+    uint256 public anotherStorageValue;
 
-    constructor(uint _storageValueToCheck)
+    constructor(uint256 _storageValueToCheck)
         public
         ContractV2(_storageValueToCheck)
     {
     }
 
-    function setAnotherStorageValue(uint _value) public {
+    function setAnotherStorageValue(uint256 _value) public {
         anotherStorageValue = _value * 2;
     }
